Fix ranking sync and manual vote handling
